CASMINST-4121 - Add container image check before removing
This is to support a 1.2 to 1.2 upgrade

+++ upgrade/1.2/scripts/upgrade/ncn-upgrade-ceph-nodes.sh
@@ -64,7 +64,10 @@ if [[ $state_recorded == "0" ]]; then
     fi
 
     ## TEMP - Remove ceph v15.2.12 from images before backups - CASMINST-4099
-    ssh ${upgrade_ncn} 'podman rmi registry.local/ceph/ceph:v15.2.12'
+    if [[ $(podman images --format json|jq '.[].Names|.[]'|grep 15.2.12) ]]
+    then
+      ssh ${upgrade_ncn} 'podman rmi registry.local/ceph/ceph:v15.2.12'
+    fi
     ## END TEMP - CASMINST-4099
 
     ssh ${upgrade_ncn} 'systemctl stop ceph.target;sleep 30;tar -zcvf /tmp/$(hostname)-ceph.tgz /var/lib/ceph /var/lib/containers /etc/ceph;systemctl start ceph.target'
